[Docs] Update 'self-hosting options' link on getting-started.mdx (#2267)
Update 'self-hosting options' link on getting-started.mdx

When browsing from: 
https://docs.twenty.com/start/getting-started/

The link with text "self-hosting options" lands on a 404 page with this address:
https://docs.twenty.com/start/getting-started/self-hosting

Adding '/start/' to the beginning of the URL fixes this. 

Note: if your browser automatically redirects from:
https://docs.twenty.com/start/getting-started/
to:
https://docs.twenty.com/start/getting-started

Then the relative document URL will work as expected. When adding the trailing slash, the link breaks.