Fix horizontal bar chart spacing between ticks (#15702)
This PR fixes a bug where the spacing between ticks on the Y axis was
adjusted according to the width of the widget where it should have been
adjusted according to the height. Same for the X axis.

**Before**:
Works for the vertical layout but not for the horizontal layout

https://github.com/user-attachments/assets/f0b94103-1ba6-4dbe-bbc6-c47541c1b5fa

**After**:
Works for both

https://github.com/user-attachments/assets/165faada-2943-4e05-92d6-a31def569500
(cherry picked from commit ea56586766a00811af70a6857bd0acc00e9c081a)